AllPara fix bug
