+++ TaskLayer/TaskLayer.csproj
@@ -94,7 +94,6 @@
     </Reference>
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="GlobalTaskLevelSettings.cs" />
     <Compile Include="EventArgs\XmlForTaskListEventArgs.cs" />
     <Compile Include="EverythingRunner\EverythingRunnerEngine.cs" />
     <Compile Include="CalibrationTask\CalibrationTask.cs" />
