+++ TaskLayer/TaskParameters/CalibrationParameters.cs
@@ -4,6 +4,12 @@ namespace EngineLayer
 {
     public class CalibrationParameters
     {
+        public CalibrationParameters()
+        {
+            PrecursorMassTolerance = new PpmTolerance(10);
+            NonLinearCalibration = true;
+            WriteIntermediateFiles = false;
+        }
         public Tolerance PrecursorMassTolerance { get; set; }
         public bool NonLinearCalibration { get; set; }
         public bool WriteIntermediateFiles { get; set; }
